# internal/apidocs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NordicIntel Storage API",
    "version": "1.0.0",
    "description": "Private current-state observation dataset storage. The health and documentation routes are public; all /v1 routes require a bearer token."
  },
...
  "paths": {
...
    "/openapi.json": {
      "get": {
        "operationId": "getOpenAPIDocument",
        "tags": [
          "Documentation"
        ],
        "security": [],
        "summary": "Public machine-readable contract",
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 document",
            "content": {
              "application/vnd.oai.openapi+json;version=3.1": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/docs": {
      "get": {
        "operationId": "redirectToDocumentation",
        "tags": [
          "Documentation"
        ],
        "security": [],
        "summary": "Redirect to the interactive documentation",
        "responses": {
          "308": {
            "description": "Permanent redirect to /docs/"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/docs/": {
      "get": {
        "operationId": "getDocumentation",
        "tags": [
          "Documentation"
        ],
        "security": [],
        "summary": "Public interactive documentation",
        "responses": {
          "200": {
            "description": "Embedded Swagger UI",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          }
        }
      }
    },
    "/v1/providers": {
      "get": {
        "operationId": "listProviders",
        "tags": [
          "Providers"
        ],
        "responses": {
          "200": {
            "description": "Providers with current datasets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "providers"
                  ],
                  "properties": {
                    "providers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Provider"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/providers/{provider}/datasets": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProviderCode"
        }
      ],
      "get": {
        "operationId": "listProviderDatasets",
        "tags": [
          "Datasets"
        ],
        "responses": {
          "200": {
            "description": "Current datasets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider_code",
                    "datasets"
                  ],
                  "properties": {
                    "provider_code": {
                      "type": "string"
                    },
                    "datasets": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DatasetSummary"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/providers/{provider}/datasets/{dataset}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProviderCode"
        },
        {
          "$ref": "#/components/parameters/DatasetCode"
        }
      ],
      "get": {
        "operationId": "getDatasetSummary",
        "tags": [
          "Datasets"
        ],
        "responses": {
          "200": {
            "description": "Dataset summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSummary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "post": {
        "operationId": "createOrReplaceDataset",
        "tags": [
          "Datasets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Replacement"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MutationResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/TooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "delete": {
        "operationId": "deleteDataset",
        "tags": [
          "Datasets"
        ],
        "responses": {
          "204": {
            "description": "Deleted or already absent"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/providers/{provider}/datasets/{dataset}/structure": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProviderCode"
        },
        {
          "$ref": "#/components/parameters/DatasetCode"
        }
      ],
      "get": {
        "operationId": "getDatasetStructure",
        "tags": [
          "Datasets"
        ],
        "responses": {
          "200": {
            "description": "Dataset structure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v1/providers/{provider}/datasets/{dataset}/data": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProviderCode"
        },
        {
          "$ref": "#/components/parameters/DatasetCode"
        },
        {
          "$ref": "#/components/parameters/Format"
        }
      ],
      "get": {
        "operationId": "getDatasetData",
        "tags": [
          "Observations"
        ],
        "responses": {
          "200": {
            "description": "Complete observation cube",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "ProviderCode": {
        "name": "provider",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 256
        }
      },
      "DatasetCode": {
        "name": "dataset",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 256
        }
      },
...
    },
    "schemas": {
...
      "Provider": {
        "type": "object",
        "required": [
          "provider_code",
          "dataset_count"
        ],
        "properties": {
          "provider_code": {
            "type": "string"
          },
          "dataset_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        },
        "additionalProperties": false
      },
      "DatasetSummary": {
        "type": "object",
        "description": "Counts and identity for the complete stored dataset.",
        "required": [
          "provider_code",
          "dataset_code",
          "source_stamp",
          "cell_count",
          "valued_cell_count",
          "null_cell_count",
          "updated_at"
        ],
        "properties": {
          "provider_code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "dataset_code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "source_stamp": {
            "$ref": "#/components/schemas/SourceStamp"
          },
          "cell_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "maximum": 1000000
          },
          "valued_cell_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 1000000
          },
          "null_cell_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0,
            "maximum": 1000000
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Provider or dataset not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected internal failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "Unsupported method for this route",
        "headers": {
          "Allow": {
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Database unavailable or operation deadline exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}